ImageClassification/python/requirements.txt
matplotlib==3.1.2
numpy==1.18.0
tensorflow==1.15.0
python-dotenv==0.10.3